+++ Scripts/Scripts.csproj
@@ -4767,12 +4767,12 @@
     <Compile Include="Services\XmlSpawner\XMLSpawner Support\StaffCloak.cs" />
     <Compile Include="SpecialSystems\CraftingTokens\CraftingTokens.cs" />
     <Compile Include="SpecialSystems\CraftingTokens\CustomCurrency.cs" />
-    <Compile Include="SpecialSystems\Engines\Imbuing\Imbuing.cs" />
-    <Compile Include="SpecialSystems\Engines\Imbuing\ImbuingB.cs" />
-    <Compile Include="SpecialSystems\Engines\Imbuing\ImbuingC.cs" />
-    <Compile Include="SpecialSystems\Engines\Imbuing\SoulForge - crafted.cs" />
-    <Compile Include="SpecialSystems\Engines\Imbuing\SoulForge - GargoyleMini.cs" />
-    <Compile Include="SpecialSystems\Engines\Imbuing\SoulForge - royalcity.cs" />
+    <Compile Include="SpecialSystems\Imbuing\Imbuing.cs" />
+    <Compile Include="SpecialSystems\Imbuing\ImbuingB.cs" />
+    <Compile Include="SpecialSystems\Imbuing\ImbuingC.cs" />
+    <Compile Include="SpecialSystems\Imbuing\SoulForge - crafted.cs" />
+    <Compile Include="SpecialSystems\Imbuing\SoulForge - GargoyleMini.cs" />
+    <Compile Include="SpecialSystems\Imbuing\SoulForge - royalcity.cs" />
     <Compile Include="SpecialSystems\MasterLooter\MasterLooterBackpack.cs" />
     <Compile Include="SpecialSystems\MasterLooter\MasterLooterContextMenus.cs" />
     <Compile Include="SpecialSystems\MasterLooter\MasterLooterGoldLedgerDeed.cs" />
